Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSS-expected.txt


PASS CSS.escape 
PASS CSS.supports, one argument form 
FAIL CSS.supports, two argument form assert_equals: CSS.supports: two argument form succeeds for custom property expected true but got false
